feat: improve existing exchange message test
